# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MediaTracker",
    "version": "0.1.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "security": [
    {
      "ApiKey": []
    },
    {
      "ApiHeader": []
    }
  ],
  "paths": {
...
    "/api/items/paginated": {
      "get": {
        "description": "Get items",
        "tags": [
          "Items"
        ],
        "operationId": "paginated",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "genres",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/MediaItemOrderBy"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SortOrder"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "onlyOnWatchlist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlySeenItems",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithNextEpisodesToWatch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithNextAiring",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithUserRating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithoutUserRating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "selectRandom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "languages",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "creators",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "publishers",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "mediaTypes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "yearMin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "yearMax",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "ratingMin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "ratingMax",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "onlyWithProgress",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "mediaType",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/MediaType"
                }
              ],
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MediaItemItemsResponse"
                      }
                    },
                    "page": {
                      "type": "number"
                    },
                    "totalPages": {
                      "type": "number"
                    },
                    "from": {
                      "type": "number"
                    },
                    "to": {
                      "type": "number"
                    },
                    "total": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "data",
                    "from",
                    "page",
                    "to",
                    "total",
                    "totalPages"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/items/facets": {
      "get": {
        "description": "Get facet counts",
        "tags": [
          "Items"
        ],
        "operationId": "facets",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "mediaType",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/MediaType"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "genres",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/MediaItemOrderBy"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "onlyOnWatchlist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlySeenItems",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithNextEpisodesToWatch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithNextAiring",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithUserRating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithoutUserRating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "languages",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "creators",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "publishers",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "mediaTypes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "yearMin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "yearMax",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "ratingMin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "ratingMax",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "onlyWithProgress",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacetsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/items": {
      "get": {
        "description": "Get items",
        "tags": [
          "Items"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "groupId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "mediaType",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/MediaType"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "genres",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/MediaItemOrderBy"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/SortOrder"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "onlyOnWatchlist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlySeenItems",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithNextEpisodesToWatch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithNextAiring",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithUserRating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "onlyWithoutUserRating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "selectRandom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "languages",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "creators",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "publishers",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "mediaTypes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "yearMin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "yearMax",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "ratingMin",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "ratingMax",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          },
          {
            "name": "onlyWithProgress",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaItemItemsResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      },
      "ApiHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "schemas": {
...
      "MediaType": {
        "enum": [
          "audiobook",
          "book",
          "movie",
          "tv",
          "video_game"
        ],
        "type": "string"
      },
...
      "MediaItemOrderBy": {
        "enum": [
          "lastAiring",
          "lastSeen",
          "mediaType",
          "nextAiring",
          "platformRecommended",
          "progress",
          "recommended",
          "releaseDate",
          "status",
          "title",
          "unseenEpisodes"
        ],
        "type": "string"
      },
      "SortOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "type": "string"
      },
...
      "FacetsResponse": {
        "type": "object",
        "properties": {
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetOption"
            }
          },
          "years": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetOption"
            }
          },
          "languages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetOption"
            }
          },
          "creators": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetOption"
            }
          },
          "publishers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetOption"
            }
          },
          "mediaTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetOption"
            }
          }
        },
        "required": [
          "creators",
          "genres",
          "languages",
          "mediaTypes",
          "publishers",
          "years"
        ]
      },
      "FacetOption": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "count",
          "value"
        ]
      },
...
    }
  }
}